fix: update character detail view for Next.js 15 async params and service integration
- Split page into server component and client component for Next.js 15 async params compatibility
- Update to use CharacterService.getCharacterById with proper ServiceResult handling
- Fix component to use ICharacter type instead of Character model
- Update tests to mock new service method structure
- Handle typed routes with appropriate casting
- Maintain TDD approach with comprehensive test coverage

Relates to #26
